Image

+++ src/HTML/FileGenerator.ts
@@ -1,5 +1,6 @@
 import * as fs from "fs";
 import * as path from "path";
+import * as vscode from "vscode";
 import { TemplateManager } from "./TemplateManager";
 import { Page } from "../../webview-ui/src/types";
 import { Node } from "./JSONParser";
@@ -12,13 +13,20 @@ export class FileGenerator {
   private projectName: string;
   private outputPath: string;
   private templateManager: TemplateManager;
+  private workspaceFolder: string;
 
-  constructor(projectName: string, outputPath: string, templateManager: TemplateManager) {
+  constructor(
+    projectName: string,
+    outputPath: string,
+    templateManager: TemplateManager,
+    workspaceFolder: string
+  ) {
     const structureGenerator = new ProjectStructureGenerator(outputPath, projectName);
     structureGenerator.generateStructure();
     this.outputPath = structureGenerator.getOutputPath();
     this.projectName = projectName;
     this.templateManager = templateManager;
+    this.workspaceFolder = workspaceFolder;
   }
 
   public generateProjectFiles(pages: Page[]) {
@@ -33,6 +41,7 @@ export class FileGenerator {
     this.createJsFile();
     this.createReadme();
     this.copyAssetImages();
+    this.copyImages();
   }
 
   private createFile(fileName: string, content: string) {
@@ -65,6 +74,7 @@ export class FileGenerator {
   private createPageFiles(pages: Page[]) {
     pages.forEach((page) => {
       resetComponentCounters();
+      console.log("Generating page content with outputPath:", this.outputPath);
       const pageContent = this.generatePageHtmlContent(page);
       resetComponentCounters();
       const gridCss = generateBackgroundCss(
@@ -100,9 +110,6 @@ export class FileGenerator {
 
   private generatePageHtmlContent(page: Page): string {
     try {
-      console.log("Generating content for page:", page.name);
-      console.log("Raw page content:", JSON.stringify(page.content, null, 2));
-
       if (typeof page.content === "string") {
         page.content = JSON.parse(page.content);
       }
@@ -152,6 +159,23 @@ export class FileGenerator {
       for (const file of assetFiles) {
         const srcPath = path.join(templateAssetsPath, file);
         const destPath = path.join(projectAssetsPath, file);
+        fs.copyFileSync(srcPath, destPath);
+      }
+    }
+  }
+  private copyImages() {
+    const sourceDir = path.join(vscode.workspace.workspaceFolders[0].uri.fsPath, "uploaded_images");
+    const destDir = path.join(this.outputPath, "images");
+
+    if (!fs.existsSync(destDir)) {
+      fs.mkdirSync(destDir, { recursive: true });
+    }
+
+    if (fs.existsSync(sourceDir)) {
+      const files = fs.readdirSync(sourceDir);
+      for (const file of files) {
+        const srcPath = path.join(sourceDir, file);
+        const destPath = path.join(destDir, file);
         fs.copyFileSync(srcPath, destPath);
       }
     }
